В логике: Исправляет некорректное задание nowTurn в LoopingList и AbstractGameState. Удаляет поле nowTurn в AbstractGameState т.к. информация о текущем игроке хранится в LoopingList. Исправляет удаление карт у игроков (удалялось у nowTurn последним действием, а nowTurn к этому моменту уже может поменяться). Изменяет api playCard, чтобы в suddenCards можно было понять у кого удалять сыгранные карты.
В сервере:
Выносит логику RoomService из GameService. Рефакторит и исправляет баги в логике работы playCard в GameService. Изменяет api взаимодействия с room, заменяя попытку сообщить user через веб сокет на тело в ответе сервера.

+++ backend/logic/src/main/java/com/kittens/logic/action/StealKnownCard.java
@@ -24,7 +24,7 @@ public class StealKnownCard implements Action
     {
         AbstractPlayer nowTurn = gameState.getNowTurn();
 
-        var playerIdWhoseCard = Integer.parseInt(playerQuestioner.ask(nowTurn, WHICH_PLAYER));
+        var playerIdWhoseCard = Long.parseLong(playerQuestioner.ask(nowTurn, WHICH_PLAYER));
         var playerWhoseCard = gameState.getPlayerById(playerIdWhoseCard);
 
         var stringTakenCardName = playerQuestioner.ask(nowTurn, WHICH_CARD_TO_TAKE);
