+++ backend/logic/src/main/java/com/kittens/logic/action/StealOfPlayerChoice.java
@@ -31,7 +31,7 @@ public class StealOfPlayerChoice implements Action
 
         AbstractPlayer nowTurn = gameState.getNowTurn();
 
-        var playerIdWhoseCard = Integer.parseInt((playerQuestioner.ask(nowTurn, WHICH_PLAYER)));
+        var playerIdWhoseCard = Long.parseLong((playerQuestioner.ask(nowTurn, WHICH_PLAYER)));
         var playerWhoseCard = gameState.getPlayerById(playerIdWhoseCard);
         
         if (playerWhoseCard.getCards().size() < 1)
